add main method back on Indexer and remove duplicated code from subclasses
