+++ lucille-core/src/main/java/com/kmwllc/lucille/indexer/CSVIndexer.java
@@ -97,5 +97,4 @@ public class CSVIndexer extends Indexer {
     }
   }
 
-
 }
